# perf: optimize scope construction and `Isolate::get_slot/set_slot()`