--- a/src/scope.rs
+++ b/src/scope.rs
@@ -645,6 +645,7 @@ macro_rules! impl_scope_drop {
     unsafe impl<$($params),*> Scope for $type {}
 
     impl<$($params),*> Drop for $type {
+      #[inline(always)]
       fn drop(&mut self) {
         data::ScopeData::get_mut(self).notify_scope_dropped();
       }
@@ -992,6 +993,7 @@ pub(crate) mod data {
     /// Returns a mutable reference to the data associated with topmost scope
     /// on the scope stack. This function does not automatically exit zombie
     /// scopes, so it might return a zombie ScopeData reference.
+    #[inline(always)]
     pub(crate) fn get_current_mut(isolate: &mut Isolate) -> &mut Self {
       let self_mut = isolate
         .get_current_scope_data()
@@ -1022,6 +1024,7 @@ pub(crate) mod data {
     ///
     /// This function panics if the root can't be activated because there are
     /// still other scopes on the stack and they're not zombies.
+    #[inline(always)]
     pub(crate) fn get_root_mut(isolate: &mut Isolate) -> &mut Self {
       let mut current_scope_data = Self::get_current_mut(isolate);
       loop {
@@ -1040,6 +1043,7 @@ pub(crate) mod data {
       isolate.set_current_scope_data(None);
     }
 
+    #[inline(always)]
     pub(super) fn new_context_scope_data<'s>(
       &'s mut self,
       context: Local<'s, Context>,
@@ -1088,12 +1092,14 @@ pub(crate) mod data {
       })
     }
 
+    #[inline(always)]
     pub(super) fn new_handle_scope_data(&mut self) -> &mut Self {
       self.new_handle_scope_data_with(|_, _, raw_context_scope| {
         debug_assert!(raw_context_scope.is_none())
       })
     }
 
+    #[inline(always)]
     pub(super) fn new_handle_scope_data_with_context(
       &mut self,
       context_ref: &Context,
@@ -1122,6 +1128,7 @@ pub(crate) mod data {
       )
     }
 
+    #[inline(always)]
     pub(super) fn new_escapable_handle_scope_data(&mut self) -> &mut Self {
       self.new_scope_data_with(|data| {
         // Note: the `raw_escape_slot` field must be initialized _before_ the
@@ -1148,6 +1155,7 @@ pub(crate) mod data {
       })
     }
 
+    #[inline(always)]
     pub(super) fn new_try_catch_data(&mut self) -> &mut Self {
       self.new_scope_data_with(|data| {
         let isolate = data.isolate;
@@ -1166,6 +1174,7 @@ pub(crate) mod data {
       })
     }
 
+    #[inline(always)]
     pub(super) fn new_callback_scope_data<'s>(
       &'s mut self,
       maybe_current_context: Option<Local<'s, Context>>,
@@ -1178,6 +1187,7 @@ pub(crate) mod data {
       })
     }
 
+    #[inline(always)]
     fn new_scope_data_with(
       &mut self,
       init_fn: impl FnOnce(&mut Self),
@@ -1213,6 +1223,7 @@ pub(crate) mod data {
 
     /// Either returns an free `Box<ScopeData>` that is available for reuse,
     /// or allocates a new one on the heap.
+    #[inline(always)]
     fn allocate_or_reuse_scope_data(&mut self) -> &mut Self {
       let self_nn = NonNull::new(self);
       match &mut self.next {
@@ -1234,6 +1245,7 @@ pub(crate) mod data {
       }
     }
 
+    #[inline(always)]
     pub(super) fn as_scope<S: Scope>(&mut self) -> S {
       assert_eq!(Layout::new::<&mut Self>(), Layout::new::<S>());
       // In debug builds, a new initialized `ScopeStatus` will have the `zombie`
@@ -1247,6 +1259,7 @@ pub(crate) mod data {
       unsafe { ptr::read(&self_nn as *const _ as *const S) }
     }
 
+    #[inline(always)]
     pub(super) fn get<S: Scope>(scope: &S) -> &Self {
       let self_mut = unsafe {
         (*(scope as *const S as *mut S as *mut NonNull<Self>)).as_mut()
@@ -1255,6 +1268,7 @@ pub(crate) mod data {
       self_mut
     }
 
+    #[inline(always)]
     pub(super) fn get_mut<S: Scope>(scope: &mut S) -> &mut Self {
       let self_mut =
         unsafe { (*(scope as *mut S as *mut NonNull<Self>)).as_mut() };
@@ -1278,6 +1292,7 @@ pub(crate) mod data {
       self
     }
 
+    #[inline(always)]
     fn try_exit_scope(mut self: &mut Self) -> &mut Self {
       loop {
         self = match self.status.get() {
@@ -1293,10 +1308,14 @@ pub(crate) mod data {
       }
     }
 
+    #[inline(always)]
     fn exit_scope(&mut self) -> &mut Self {
       // Clear out the scope type specific data field. None of the other fields
       // have a destructor, and there's no need to do any cleanup on them.
-      self.scope_type_specific_data = Default::default();
+      if !matches!(self.scope_type_specific_data, ScopeTypeSpecificData::None) {
+        self.scope_type_specific_data = Default::default();
+      }
+
       // Change the ScopeData's status field from 'Current' to 'Free', which
       // means that it is not associated with a scope and can be reused.
       self.status.set(ScopeStatus::Free);
@@ -1306,6 +1325,7 @@ pub(crate) mod data {
       self
         .get_isolate_mut()
         .set_current_scope_data(Some(previous_nn));
+
       // Update the parent scope's status field to reflect that it is now
       // 'Current' again an no longer 'Shadowed'.
       let previous_mut = unsafe { &mut *previous_nn.as_ptr() };
@@ -1336,6 +1356,7 @@ pub(crate) mod data {
     /// because the 'a lifetime ends there.
     ///
     /// Scope types that do no store local handles are exited immediately.
+    #[inline(always)]
     pub(super) fn notify_scope_dropped(&mut self) {
       match &self.scope_type_specific_data {
         ScopeTypeSpecificData::HandleScope { .. }
@@ -1355,18 +1376,22 @@ pub(crate) mod data {
       }
     }
 
+    #[inline(always)]
     pub(crate) fn get_isolate(&self) -> &Isolate {
       unsafe { self.isolate.as_ref() }
     }
 
+    #[inline(always)]
     pub(crate) fn get_isolate_mut(&mut self) -> &mut Isolate {
       unsafe { self.isolate.as_mut() }
     }
 
+    #[inline(always)]
     pub(crate) fn get_isolate_ptr(&self) -> *mut Isolate {
       self.isolate.as_ptr()
     }
 
+    #[inline(always)]
     pub(crate) fn get_current_context(&self) -> *const Context {
       // To avoid creating a new Local every time `get_current_context() is
       // called, the current context is usually cached in the `context` field.
@@ -1390,6 +1415,7 @@ pub(crate) mod data {
       }
     }
 
+    #[inline(always)]
     pub(super) fn get_escape_slot_mut(
       &mut self,
     ) -> Option<&mut Option<raw::EscapeSlot>> {
@@ -1399,6 +1425,7 @@ pub(crate) mod data {
         .map(|escape_slot_nn| unsafe { escape_slot_nn.as_mut() })
     }
 
+    #[inline(always)]
     pub(super) fn get_try_catch(&self) -> &raw::TryCatch {
       self
         .try_catch
@@ -1407,6 +1434,7 @@ pub(crate) mod data {
         .unwrap()
     }
 
+    #[inline(always)]
     pub(super) fn get_try_catch_mut(&mut self) -> &mut raw::TryCatch {
       self
         .try_catch
@@ -1420,6 +1448,7 @@ pub(crate) mod data {
     /// default values. This function exists solely because it turns out that
     /// Rust doesn't optimize `Box::new(Self{ .. })` very well (a.k.a. not at
     /// all) in this case, which is why `std::alloc::alloc()` is used directly.
+    #[cold]
     fn boxed(isolate: NonNull<Isolate>) -> Box<Self> {
       unsafe {
         #[allow(clippy::cast_ptr_alignment)]
@@ -1481,6 +1510,7 @@ pub(crate) mod data {
   }
 
   impl Drop for ScopeTypeSpecificData {
+    #[inline(always)]
     fn drop(&mut self) {
       // For `HandleScope`s that also enter a `Context`, drop order matters. The
       // context is stored in a `Local` handle, which is allocated in this
@@ -1538,6 +1568,7 @@ mod raw {
   }
 
   impl Drop for ContextScope {
+    #[inline(always)]
     fn drop(&mut self) {
       unsafe { v8__Context__Exit(self.entered_context.as_ptr()) };
     }
@@ -1566,6 +1597,7 @@ mod raw {
   }
 
   impl Drop for HandleScope {
+    #[inline(always)]
     fn drop(&mut self) {
       unsafe { v8__HandleScope__DESTRUCT(self) };
     }
@@ -1624,6 +1656,7 @@ mod raw {
   }
 
   impl Drop for TryCatch {
+    #[inline(always)]
     fn drop(&mut self) {
       unsafe { v8__TryCatch__DESTRUCT(self) };
     }