Before (on a MacBook Pro M1 2021):

```
Running function_overhead ...
  18.0 ns per run 55.5 million ops/sec → new_
  18.4 ns per run 54.5 million ops/sec → new_raw
  13.5 ns per run 74.1 million ops/sec → new_set_uint32
  8.8 ns per run 113.4 million ops/sec → new_raw_set_uint32
  2.7 ns per run 366.3 million ops/sec → new_fast
Running primitives ...
  15.0 ns per run 66.7 million ops/sec → undefined_from_scope
  7.6 ns per run 131.8 million ops/sec → undefined_from_isolate
```

After:

```
Running function_overhead ...
  15.1 ns per run 66.2 million ops/sec → new_
  15.1 ns per run 66.1 million ops/sec → new_raw
  11.5 ns per run 87.2 million ops/sec → new_set_uint32
  6.8 ns per run 146.6 million ops/sec → new_raw_set_uint32
  2.7 ns per run 367.6 million ops/sec → new_fast
Running primitives ...
  12.2 ns per run 81.9 million ops/sec → undefined_from_scope
  7.6 ns per run 131.9 million ops/sec → undefined_from_isolate
```